EmptyReturnFixer - move fixer outside of Symfony level
